NBA map function

NBA maps for protection level, threat status, and condition. The function can take a dataset with either one polygon or multipolygon per ecosystem (each row has a unique ecosystem) or multiple rows per ecosystem type.

Usage

NBA_map(DF, COLS = NULL, GEOM, CAP, FILL)

Arguments

DF

The data frame that contains the information

COLS

The variables to group together with the ecosystem types if the data isn't already in the form of one row per ecosystem type.

GEOM

The name of the geometry column

CAP

The caption for the map (will be place at the bottom left of the plot)

FILL

The column that contains the categories to colour the ecosystems by (e.g. protection level, threat status, or condition)

Value

Returns a map

Details

The function expects an sf object.

Please look at the example data NBA_example_map_data and NBA_example_donut_plot to see the correct structure for the data.

The name of the fill column is irrelevant, but the categories must be spelled correctly (there is a list of the standard spellings/ cases of NBA categories named NBA_categories in this package, which can be accessed for reference).

Examples

#map <- NBA_map(
#DF = mem_2023,
#COLS = c(ecosystem_type, protection_level_30, threat_status, eco_poly_colour),
#GEOM = geometry,
#CAP = "Figure 1. Map of the distribution of the ecosystems threat status",
#FILL = threat_status)

#map